+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+* New option `--path` to set the path to the ROMs folder. By default it's `/home/pi/RetroPie/roms`.
 * Warning dialog when using the `--convert-all` option saying that it can take a lot of time to convert all the videos.
 
 ## [2.0.0] - 2019-04-15